Rename helpers.ts to index.ts
