+++ src/classes/Projects.ts
@@ -1,6 +1,6 @@
 import { ProjectTag } from "@/constants/index";
 import { ValidLinkHref } from "@/elements/Link";
-import { extractGitHubUserRepo } from "@/helpers/helpers";
+import { extractGitHubUserRepo } from "@/helpers";
 import projectRoutes from "@/projectRoutes";
 import { RouteRecord } from "vite-react-ssg";
 interface ProjectClass {
